refactor(ts-plugin): move test/builder.ts under src/test/ to match other packages
